requirements.txt
absl-py==2.0.0
addict==2.2.1
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
alabaster==0.7.13
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.3
cmake==3.27.6
comm==0.1.4
conch @ git+https://github.com/Mahmoodlab/CONCH.git@32697323df54da0704aac88acee4833d202ff5ee
contourpy==1.1.1
cycler==0.12.0
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dicomweb-client==0.59.1
dill==0.3.7
docker-pycreds==0.4.0
docutils==0.19
einops==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.19
fastjsonschema==2.18.1
filelock==3.12.4
Flask==2.3.2
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
Glymur==0.12.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
h5py==3.9.0
histolab==0.6.0
huggingface-hub==0.17.3
idna==3.4
imagecodecs==2023.3.16
imageio==2.31.5
imagesize==1.4.1
imgaug==0.2.6
importlib-metadata==6.8.0
iniconfig==2.0.0
# install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lit==17.0.1
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
mlxtend==0.23.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==6.5.6
notebook_shim==0.2.3
numba==0.57.0
numcodecs==0.11.0
# numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nystrom-attention==0.0.9
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
openslide-python==1.2.0
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydicom==2.3.1
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.2.3
pytorch-toolbelt==0.6.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
referencing==0.30.2
regex==2023.10.3
requests==2.30.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
# scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.13.1
Send2Trash==1.8.2
sentry-sdk==1.34.0
setproctitle==1.3.3
shapely==2.0.1
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.17.1
threadpoolctl==3.2.0
# tiatoolbox==1.4.1
tifffile==2023.4.12
timm==0.9.8
# timm-ctp @ file:///autofs/homes/007/tg020/mmodal-xai-brca-path/code/mizero/assets/timm_ctp.tar
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.2.2
torch_geometric==2.5.0
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.17.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.1
transformers==4.31.0
triton==2.2.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.3
uri-template==1.3.0
urllib3==2.0.6
wandb==0.16.0
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
wsidicom==0.9.0
xxhash==3.4.1
y-py==0.6.0
yarl==1.9.2
ypy-websocket==0.8.4
zarr==2.14.2
zipp==3.17.0
